Se aniadio un boton de buscar en la parte de Farmacias y se cambio el layout de algunas pantallas

+++ app/src/main/java/com/example/proyecto_progra_3/Producto.kt
@@ -1,7 +1,7 @@
 package com.example.proyecto_progra_3
 
-class Producto (val nombre:String, var precio: Float, var cantidad: Int){
-    fun cambiarPrecio(nuevoPrecio:Float){
+class Producto (val nombre:String, var precio: Double, var cantidad: Int){
+    fun cambiarPrecio(nuevoPrecio: Double){
         precio = nuevoPrecio
     }
